Decouple transparent and erased properties
It is at times useful to have `transparent` methods that are not
`erased`. For instance, a transparent `unapply` might enable inline
reductions, but we want to also keep it at runtime, because the same
type might be matched at compile-time and at run-time.

This commit decouples `transparent` and `erased` again. `transparent`
does not imply `erased`.