Remove class tags from Tasty reflect interface
We found that when the implementation of those class tags
uses the same class tags, type matches are unsound.
Instead, we replace the more precise scrutinee by equivalent
extractors. In the future, we intend to  language support
for abstract type pattern matching where the scrutinee
can get a more precise type.