Change typing of "throw" expressions to allow a nullable argument
The JLS (https://docs.oracle.com/javase/specs/jls/se7/html/jls-14.html#jls-14.18)
explains that "throw" should accept a null argument (in which case it
throws an NPE):

```
If evaluation of the Expression completes normally, producing a null value, then
an instance V' of class NullPointerException is created and thrown instead of null.
The throw statement then completes abruptly, the reason being a throw with value V'.
```

So changed the typer so the prototype allows for null.