Stop using the bootclasspath for Java 9 compatibility
It's no longer introspectable under Java 9. The scalac 2.12 shell scripts work
around this by passing the bootclasspath jars as an additional system
property, but we can avoid this complication by just using -classpath instead.
This may have a slight negative impact on startup performance since the
JVM bytecode verifier is not run for classes from the bootclasspath, but
we can worry about that later.