Use a special type for match aliases
This is a more principled implementation of the situation that
match aliases are not full type aliases because one side cannot
be freely substituted for the other. Match aliases are checked
in some respects like abstract type bounds (in particular,
nested cycles are allowed), but are treated as aliases in others
(in particular, lower and upper "bound" are always guaranteed
to be the same).