Add NonEmptyTuple abstract class
... and move `_(_)`, `_.head`, `_.tail` to it. That way,
we no not need to specify a specific (probably wildcarded) pair type
when invoking these operations.

I first tried to move the three operations simply to Tuple. But since
`Unit` is a `Tuple` as well, this gives us funny error messages for `apply`. For instance
this one in errMsgTests:

```
object Scope{
  def foo(a: Int) = ()
  foo(1)("2")
}
```
If `Tuple` had an `apply` method, this would give `expected: Int, found : String`