Handle JavaNull when converting varargs
When the compiler encounters a method with a varargs argument, the type
of the varargs is initially represnted as an Array[T]. Later, it is
transformed into a RepeatedParamType[T].

However, the nullability transform makes it so that the varargs has type
`Array[T|JavaNull]|JavaNull`. We need to teach `arrayToRepeated` how to
handle that case so we can get `RepeatedParamType[T|JavaNull]|JavaNull`
as the result.