IDE tests: compile dependent workspaces
When writing IDE tests that simulate multi-project setups, we need to
compile the sources of the workspaces that are depended on, because the
dependent workspaces require their class files.

This commit changes the `TestServer` so that workspaces that are
depended on are compiled during the initialization of the server. When
there are no dependency relation between workspaces, no compilation is
performed.