Pass empty string instead of null in dummy constructor
When an enum is read from Java code, the compiler synthesizes a bunch
of classes/modules/fields for it.

One of the synthesized entities is a class that extends java.lang.Enum
and calls its constructor. The first argument of the constructor is a
string, so we were passing null which failed. Pass the empty string
instead. This is ok because the synthesized Java code isn't run: it's
just there for typechecking.