Equate TupleN(...) and *: types
Make `T1 *: ... *: Tn *: Unit` a subtype of `(T1, ..., Tn)`. This is sound since
the two types are erased to the same representation.

I tried alternatively to (...) types to `*:` types instead of `Tuple` types.
But then we have to translate `TupleN` classes as well since these can come
from Scala-2. This looked more fragile than the solution in this commit.