Handle constructors properly in nullability transform
This adds a second TypeMap, specifically for constructors.
For constructors, we nullify the argument types, but not the return
type.

Once we nullify the arguments of constructors, all case classes were
breaking, because the logic to generate synthetic methods for case
classes relies on finding the symbol for IndexOutOfBoundException, which
changed with this CL. Patch up that logic as well.