Feature Removal: Do not check call-site @tailrec anymore.
This feature does not exist in Scala 2, and was never documented in
Scala 3. Its only internal use comes from a swiping commit that
added as many `@tailrec` annotations as possible all over the
codebase: see be5720c18ca6768c7e72d4258677952848db2bb4.

Supporting that feature became problematic when TailRec was moved
after erasure, because we needed an attachment to remember
call-site `@tailrec` annotations, which would otherwise be removed
by erasure.

This commit removes the feature, on the grounds that its
cost/benefit is too high.