Un-deprecate linesIterator
Java 11 introduces the `lines` method on `String`, which means we should
probably avoid using that name, and go back to `linesIterator`.

To allow compiling with -Xfatal-warnings on Java 11, we have no choice
on 2.12 but to un-deprecate this method. Probably shouldn't immediately
deprecate `lines`, since most users will be running on Java 8, and are
thus not affected by this. Perhaps 2.13 is the right time frame for
switching around the deprecation.