sbt-dotty: Make launchIDE work offline
For each Code extension we require, we start by trying to install or
upgrade the extension. If this fails we check if an existing version of
the extension exists. If this also fails we throw an exception. This
ensures that we're always running the latest version of the extension
but that we can still work offline.