Drop incorrect optimization in substituters
The optimization was to disregard a prefix of a named type if its symbol is static
and none of the mapped symbols in `from` is static. This is incorrect, since the
prefix might be a non-static reference to a static object. In that case the prefix
still has to be considered. i2895a.scala contains code that failed when running
with the cbn implementation of `assert` in the next commit.