Fix `go to definition` with overridden methods
When doing go to definition on a reference, we are supposed to look for
the symbol under the cursor, and the symbols that override it in the
same source file. We were previously looking only in the tree of the
first definition, and would not inspect the rest of the sourcefile.