Fix #5242: Don't run all Scala files as worksheets
When initializing the `WorksheetRunFeature`, we were specifying that the
feature should apply to all files where the `document selector` matches
that of the Dotty plugin, or files in the Scala language whose name ends
in `.sc`. Unfortunately, this matches all `.scala` files, and these were
run as worksheets.

This commit changes the document selector for the `WorksheetRunFeature`
so that Scala files are not run as if they were worksheets.