A mutable HashMap that degrades to red-black trees in the worst case
The implementation is based on my new HashMap. Hash buckets use linked
lists ordered by hashcode by default. When a bucket gets too big due to
hash collisions it is converted into a red-black tree based on an
implicit Ordering for the keys. This is similar to java.util.HashMap
except we are using an Ordering instead of requiring the key type to
implement Comparable.

Equality semantics are determined by the Ordering, which has to be
consistent with the key’s hashCode.