IDE: Find and add workspace root if there are none
When the IDE is opened without a workspace (no folder is open), and then
a Scala source is opened, we will now search for the closest possible
workspace root (the closest parent directory that contains a
`build.sbt`). If no such directory can be found, the parent directory of
the file that has been opened is chosen. This folder is then added as a
workspace folder.

Adding a new workspace folder triggers a reloading of the extension. The
code of the extension will be executed in an environment  where a
workspace root is set, and Dotty IDE will be able to start correctly.

Fixes #5281