Remove Worksheet dependency on JLine
The worksheet used to spawn a REPL with JLine just to be able to read
from an InputStream.

This is alternative solution that simply uses java.util.Scanner to read
from an InputStream. For that, we need to agree on a delimiter. We
choose a character within the private user area to not conflict with
the output produced by the REPL.