Fix #5372: Use `applyDynamic` in structural calls
This commit changes how `Selectable` works for calling methods: It used
to be called `selectMethodDynamic and to receive the name of the method
to call along with the class tags of the formal parameters.

This commit renames it to `applyDynamic`, and changes it so that it also
takes the actual arguments that are passed in the function call.

Fixes #5372