Improve display of documentation in IDE
This commit introduces `ParsedComment` which is used to parse the doc
comments and make it easier to retrieve parts of the comments (for
instance, the documentation associated with a given method parameter).

The documentation marked with the tags that scaladoc supports are
extracted and their content are formatted into lists, code fences, etc.