Fix #5140: Ycheck failure in covariant java varargs
Allow passing `Array[Dog]` to a Java varargs that expects an
`Array[Animal]`. This types because Java vargs are represented as
`RepeatedParam[+T]`, which is covariant.

However, after `elimRepeated`, the `RP`s go away and the true type
(Array) is revealed, which was causing a Ycheck error because arrays are
invariant before erasure.

Fix the Ycheck error by changing how `elimRepeated` transforms `RP[T]`s that
appear in Java code.

The previous transform was `RP[T]` => `Array[T]`.
The new transform is `RP[T]` => `Array[_ <: T]`.

This is unsound but consistent with the typer behaviour and what scalac does.