Don't widen union types of the form `T|Null`
There are two places where union types are widened:
1) In the r.h.s of member definitions
e.g.

val x = foo() // foo returns a union

2) When inferring a type parameter

foo(if (1 == 2) "hello" else 4) // type parameter inferred to be Any

In both of these cases, avoid widening the union if one of its
constituent types is `Null`.

The main reason for this is that the widening `T|Null` is likely to
result in `Any`, which isn't a very useful type.

Conversly, it's useful to not lose information about the nullability of
a type.