Make local copy of opaque type inaccessible for the outside
i5455.scala showed another misuse: One could get at the local copy
of an opaque type even though this type was not defined. This commit
closes that loophole.

Aside: It might be better to do the opaque type expansion in Namer
instead of in Desugar. This could let us use a synthetic name for
the local alias and thereby avoid namespace pollution. Right now one cannot
define a type with the same name as the opaque type alias in the
latter's companion object. It's a crazy thing to do, but still...