Report the correct errors in the IDE
In some situations (like when interpolating type variables) the logic
changes depending on whether we have unreported errors or not.
Before this commit, we assumed that any error in a StoreReporter was
unreported, but this is wrong: the IDE, the REPL and other tools use
a StoreReporter as their root reporter.

This commit fixes this by making sure all the code that looks for
unreported errors does so by calling `hasUnreportedErrors` and
making that method return false when there is no outer reporter.