Properly type bounds of wildcards in hk positions
Typing a TypeBoundsTree(Empty, Empty) leads to a TypeBoundsTree(Nothing,
Any), if this tree appears in a higher-kinded position, we manually set
its type, but we don't change the upper bound which is still "Any" and
ends up being pickled. To avoid this issue this commit simply replaces
the TypeBoundsTree with a synthetic TypeTree with the proper type and
position, which is all that we need.