Integrate accumulators into collection hierarchy
Accumulator extends Iterable, but not Seq. The reason is that they
can contain more than `Int.MaxValue` elements, so we cannot provide
a reasonable implementation for `length`. This would break operations
that depend on `length` in their implementation.

Instead of the `.accumulator` extension methods in java8-compat,
collections and arrays can be converted to accumulators by calling
`.to(Accumulator)`. The hand-specializations (`IntAccumulator`) are
created with `.to(IntAccumulator)`, or with
`.to(SpecializedAccumulator)` (maybe we should make the `Accumulator`
companion the version that automatically picks the specialized version).

TODO: check Accumulator implementations, maybe some of its methods are
redundant now.