Insert fake self definition when in IDE
When parsing the source with mode = `Interactive`, insert fake self
definitions in templates.

For the IDE, The end goal is to be able to regroup `this` and self
references under a same term symbol.

Currently, a self definition has a symbol. The symbol of references to
the self definition is the enclosing class symbol. The symbol of `this`
is the enclosing class symbol.