toScala extension method for java streams
The `toScala` method takes a factory, just like the `to` method in
Scala collections.

Special handling for converting to accumulators:
  - `(s: Stream[T]).toScala(Accumulator)` creates a primitive
    accumulator (`IntAccumulator`) if `T` is primitive
  - parallel streams are converted to accumulators in parallel
  - parallel streams are always converted to accumulators (in parallel)
    before the final conversion to the target Scala collection
  - sequential streams are converted lazily to lazy collections

This PR also changes the `to` method of accumulators to support lazy
collections (use `factory.from` instead of a builder). The iterator
is based on the `AccumulatorStepper`. TODO: check if the there's a
performance penalty to using that iterator, if an AccumulatorIterator
would be faster.