Don't generate fallback Eq in Null case
Given `x == null`, to decide whether to generate the fallback
`Eq` instance, we would
  - check if x's type and Null can be assumed to be equal via
`assumedCanEqual`
  - if not, then check whether neither of them has its own `Eq` instance

However, since `Null` is no longer a subtype of other types, it no
longer has any already-defined Eq instances. This means that step 2
above succeeds, meaning that we could compare T <: AnyVal with Null,
which (I think) we don't want to do.

The fix is to add an extra check between the two above to only allow
comparisons to Null if `assumedCanEqual` would've returned true (which
only happens for reference types).