Rewrite `typer.TermRefSet`
Rewrite motivated by 2.13 collection rewrite.

`TermRefSet` used to extends `mutable.Traversable[TermRef]`
which is now deprecated. If we want to keep with the spirit
of the previous API, the right thing would be to replace
`Traversable` with `Iterable`. But then we'd have to
implement an iterator, which might be less efficient.