Better error messages for missing symbols
When I changed branches, the compiler encountered a missing required symbol
"DottyPredef.ImplicitConverter" because in the orevious branch that symbol
had been renamed to "ImplicitConversion". The error message was not exactly
helpful, however:

   None of the required alternatives of <none> satisfies required predicate

This commit fixes this. It gives a meaningful error messages in all cases where
a required symbol is missing.