Make nextTreeId counter a context field
This removes some technical debt: nextId was a global counter, so
unreliable under parallelism. That was OK so far since it was only
intended for debugging. But now we want to give more significance to
tree ids, since they will be used to determine the source file of
a tree. Hence, we need to put this aspect in order by making nextTreeId
a context field. This entailed a lot of changes, so no wonder it was
not done before.