Make Modifiers not extend from Positioned
Modifiers was made a subclass of Positioned since it contains elements that
are Positioned: source modifiers and annotations. That way, we got the navigation
logic to get to the elements for free. However, a lot of modifiers are synthetic,
and seems artificial to give these a position. With the addition of a source
component to position the balance swings the other way. It now seems preferable
to have Modifiers not extend Positions and to consider modifier as a as a separate
case for navigation instead.