Fix #5666: Fix normalizing map when testing for seen pairs
When testing for seen pairs we subject both types to a "normalize" type map
that derefereces LayVals. This is tricky but necessary. But we should take
care that the map does not disturbe the types in any other way. In particular
it should not replace a TypeVar by its instance.