[backport] 11330: Port fix for byte array hashcode
Issue 10690 fixed this issue for scala 2.13, and a similar
fix has been implemented here. There are some additional
changes over those in 10690, removing byte array hashing
entirely. Hashing is now consistent in that it produces
the same result regardless of the underlying numeric type,
in so far as the numbers in question fit into the byte
width of the type.