Define `reflect.Manifest` as an `object` rather than a `val`.
Since there is a `trait Manifest`, the term of the same name should
be in the same file. Otherwise, in some weird circumstances, like
somehow while building Scala.js in the community build (but only in
the community build), we get a compile error saying that `Manifest`
is already defined in the package object.

Since we cannot define a top-level `val`, we declare it as an
`object` instead. This means that we have to explicitly forward all
the public members of `ManifestFactory` in `Manifest`.

This is meant as a small incremental fix that does not impact the
compiler (since its codegen refers to `ManifestFactory`).
Longer term, we should make the codegen use the `object Manifest`
instead, and eventually retire `ManifestFactory`.