Avoid direct Skolem types in lower GADT bounds
By constraining T s.t. T >: Sko(U) and T >: Sko(U), we end up with
T >: Sko(U) | Sko(U), which is a singleton type union and is
simplified to T >: U, which is wrong for this case.

To avoid this, we record only the _consequence_ of having a Skolem type
in a bound.

This is a temporary workaround while we cannot take unions and
intersections of singleton types and should be removed in the future.