Full bootstrap: Use Dotty as the reference compiler
The non-bootstrapped compiler is now compiled by Dotty 0.14.0-RC1
instead of Scala 2.12.8. The bootstrapped compiler is compiled using the
non-bootstrapped compiler as before.

For now, we still support compilation with Scala 2.12, this is achieved
by having separate `src-2.x` and `src-3.x` directory for the library.

To use Scala 2.12 as the reference compiler, run:

++2.12.8

in sbt, or change `referenceVersion` to "2.12.8" in Build.scala.