Avoid cycles when unpickling the stdlib from TASTY
`adjustForTuple` shouldn't be used when unpickling Java (because the
Tuple classes are always defined in .scala files) nor when unpickling
TASTY (because at that point, `adjustForTuple` has already been called
by Namer before pickling). This avoids some cycles when unpickling
the standard library from TASTY (we don't have a published
Dotty-compiled standard library yet, but trying to use incremental
compilation in the stdlib that is part of our community build is enough
to trigger the issue).

Note: This won't handle the hypothetical future case of Scala 2
generated TASTY files which wouldn't have the extra parents, but let's not
worry about that for now.