Fix cyclic reference in Trees.scala
The previous commit added _root_.scala.Serializable as a parent to every
object, this lead to a cyclic reference manifesting itselfs as "not
found: type Type" in Trees.scala that we worked around by replace `Type`
with `Types.Type`. This commit fixes this properly by adding a shortcut
in typedIdent when typing `_root_` which avoids forcing imports.