Use <:< for Option.unzip
Currently, it uses asPair: A => (A, A).

The type checking error it produces is:

    scala> Some(42).unzip
                    ^
    error: No implicit view available from Int => (A1, A2).

It should be asPair: A <:< (A, A), and give the error:

    scala> Some(42).unzip
                    ^
    error: Cannot prove that Int <:< (A1, A2).

The <:< symbol is cryptic, but the error overall is much better than
one that mentions "views".  This also makes unzip consistent with the
rest of the library, which uses <:<, see orNull and flatten.