Fix #5809: Run the community-build on every PR
This commit is meant to replace the infrastructure previously setup in
https://github.com/lampepfl/dotty-community-build. A new sbt project,
community-build, is added to dotty's Build.scala and contains all the
integration tests for the community build.

In case of failure the tests will output instructions on how to
reproduce the error without JUnit. Hopefully these instructions are
sufficient to have everyone fix regressions as they are introduced,
or as a very last resort, update the code for the community build.

This should be the main improvement over the previous approach, where
the community build maintenance was centralized (Allan did all the work)
and asynchronous (it's now broken for more than a month...).

Another improvement is that a published local version of the sbt plugin
is used when running the integration test, so breakage to sbt-dotty
should be caught earlier than before.