Clean up TypeVar insertion/removal in SmartGADTMap
If we do not insert TypeVars into the bounds every time, then the only
time we need to remove them is when taking the full bounds of some type.

Since that logic now resides in ConstraintHandling and replaces all
TypeParamRefs internal to SmartGADTMap, we have no need to perform
expensive type traversals.

This removes the only reason for caching bounds.