worksheet/publishOutput returns a range instead of line
This command is sent from the server to the client when a statement in
the current worksheet has finished running, but statements can be
multi-line so returning just the line number loses some information.

We use this extra information in vscode-dotty so that the hover pop-up
containing the output of the statement is displayed when hovering over
any part of the statement, and not just the last line. To avoid breaking
this hover pop-up when using an old version of Dotty, some compatibiltiy
code is necessary.