Future-proof the collections for Dotty
Just like what we did for concat in
https://github.com/scala/scala/pull/7696, we add a dummy parameter list
to avoid signature clashes between mixin forwarders, but this change
should be less controversial since it only affects a deprecated method.

Note that this isn't actually needed to compile with Dotty master yet
because we only emit mixin forwarders when required to, but we will soon
switch to emitting them all the time like scalac
does (https://github.com/lampepfl/dotty/issues/2563), therefore we need
to get the fix in first (and most importantly in time for 2.13).