Take implicit parameters into account for is-as-specific computations
Number of implicit parameters is applied as a final tie break, if
two alternatives would be otherwise equally specific.

I tried to roll them into type specificity first, but this fails
the dotc bootstrap where we have:

     def name(implicit ctx: Context): Name  in class Denotation

     val name: Name                         in subclass SymDenotation

The `val name` wins, since it is defined in a subclass and the result types
are the same. If we let the type `(implicit ctx: Context): Name` win over
`Name` this would be ambiguous instead.