Delete DottyByteCodeTest
I am giving up. This is the worst sort of blindly testing the status quo! It tests some random
code and if that code has changed (because there is some optimization applied now) it fails.
Great! Problem is there's no way to automatically update the code. One has to go in manually and
fix the expected program including all labels! Copy pasting from the diff does not work and there's no
--update-check.

As long as things are like this I am afraid that bytecode tests are throwaway tests. They
blindly test the status quo, and if that changes our best action is to throw away the test.