Normalize infix operations
Everywhere: put eq/ne in backticks.

In dotc/core: Fix also other uses of alphanumeric uses to correspond to
new infix rules.

Question: Why not make eq/ne @infix?

`eq`/`ne` are abbreviations, not words. `min`, `max`, `to` `until`, `by` are words, so it makes
sense to use them infix like this. But for `eq` and `ne` it feels better to mark them explicitly
as an operator, which is achieved by the backticks.