Avoid non-determinism in stdlib test by sorting files
For some mysterious reason I still need to investigate, some invalid
super-calls in the stdlib that should be detected by the previous commit
are only detected when the files are compiled in a specific order, this
is made more confusing by the order of listed files being
non-deterministic by default. As a first step towards sanity, sort the
files to avoid non-determinism and consistently detect the invalid
super-calls.