Normalized accumulators to build like other collections.
Note that Accumulator itself cannot be a standard IterableFactory because it
has to dispatch to different implementation classes.  However, it now has
duplicates of all the construction methods (fill, tabulate, etc.).

It would be nice to have a type-aware dispatch to an appropriate factory
so you could `Accumulator.of[Int].empty` and have the collection type
inferred.