Move i3430.check to pending
The precise error message i3430.scala depends on the order in which files
are found. This is OK. But unfortunately it means there's no easy way to
write a check file. My (local) testCompilation, CI test, and CI bootstrapped
tests all disagree pairwise which two implicit instances are reported as
ambiguous. We could invest in a more specialized test that just checks
that an ambiguity is reported no matter what the participants. But my
patience has run out, so it would be good if someone else who picks
this up could do it.