Fix typing of new
Assume
```
   class C[X]
   type A[X] = C[X]
```
The previous rule for `new A` overwrote the type of `A`
with `C`, in order to satisfy the condition that the type
of `new` should be a class type. Overwriting types like
that in Typer is wrong, since it is not preserved by pickling.
We tried to patch up things in RefinedPrinter, but this is not
enough, as test case `alias-new.scala` shows. This test caused
the type in a `new` to be unpickled as a <notype>.

The fix is to treat eta-reduction in new types in a purely functional
way, applied in both Typer and TypeAssigner.

The problem was encountered when systematically testing creation methods,
but it has nothing to do with them.