Fix classpath issues with the dotty-sbt-bridge build
When compiling and running the tests for the bridge, we should have the
bootstrapped library on the classpath since this is what people will
actually use in the end. This is tricky to do since the bridge itself
cannot be compiled with anything bootstrapped on the classpath since the
bootstrapped projects need the bridge to be compiled !

We sidestep this issue by using two separate projects for the bridge and
its tests.