Fix test
The test previously compiled by accident due to the "more implicit parameters win" rule.
Once that rule got changed it should have failed, but it seems the error (illegal
forward reference) was not detected. Rebasing this commit on latest master detects
the error:
```
  implicit val co_i: Conversion[Char, Position[CharSequence]] = the[Conversion[Char, Position[CharSequence]]]
                                                                                                             ^
                                             co_i is a forward reference extending over the definition of co_i
```
The fix to the test avoids the error again.