Split quoted.Expr casting from sealing
Previously `seal` always casted the `quoted.Expr[T]` for some given `T`. Now, `seal` and `cast` can be used independently. If we need to seal to an `Expr[Any]` there is no need to check the that the type conforms. This use-case is common enough to deserve the optimization.

Additionally, due to bootstrapping issues, it is not practical to seal to a `quoted.Expr[Any]` within the library as we cannot create a `quoted.Type[Any]`.