Fix companion parents for enum case
In the case where an enum case contains an extends clause whose type is not
fully specified, we cannot generate a function type parent for the companion
object.

In fact, this business of generating function types as parents of case class
companions is getting more and more fragile. We should maybe get out of it
altogether.