Fix #5897 and fix #6200: implicits not retained in REPL
When evaluating definitions in REPL (`def`, `implied` etc),
they are brought in scope for all the subsequent evaluations of the
user input. This is done via the `import` mechanism. However, in Dotty,
the semantics of the imports changed: by default, the implied values are
not imported. This commit makes sure that all the members, including the
implied ones, are imported from the previous evaluations of the input.