Remove CHANGELOG.md
It complicates the release process. In order to generate CHANGELOG,
we need to tag a commit to be RC1. However, since CHANGELOG belongs
to the release candidate RC1, in order to tag a commit as RC1,
we must first include the CHANGELOG into that commit. Hence it is best
to remove this cyclical dependency altogether and keep the release
notes at the GitHub releases page.