Do not rename names starting with `$` in quoted patterns
Previously we renamed `val $a` to `val a` and at the same time created a binding
named `a`. A reference to `a` inside a splice could not be disambiguated.

```scala
case '{ val $a: Int = 4; `$a` + ${ ... a ... '{ .. `$a` .. } ... } } => a
```

This change allows us to both refer to the `a` binding and the `$a` reference inside splices without ambiguities.