Refine purity checking for constants
A pure expression such as `(a + b).c` of constant type is now characterized
again as a pure path (since it will be eventually replaced by that constant,
which is a path).

This fixes instability problems in pickling and improves inlining. For instance
`assert(0 == 0)` is reduced to `()` again with this change.