Fix interaction between by-name implicits and inlining
The construction of the dictionaries needed to support recursive by-name
implicit arguments requires trees to be hoisted out of the arguments as
they are elaborated. Moving trees around in this way has always been a
delicate operation because of the need to maintain correct owner chains.

Inlining and the position checking phase added in 69cf22021c complicates
this still further because we now have to make the source associated
with the hoisted tree match the checkers expectations after it has been
moved. This is done by reusing the repositioning logic from the inliner.

The issue this commit fixes only shows up in cases where an inline
method defined is used as part of a by-name implicit argument resolution
in a different source file.