Fix #6450: Don't drop apply alternatives in overloading resolution
The overloading resolution algorithm dropped alternatives reachable through applys
when testing for applicability. It did so in three situations

 - when comparing the arities of alternatives
 - when testing via isDirectlyApplicable
 - when testing via isApplicable

The third problem stemmed from confusion between overloaded versions
of `isApplicable` (it's ironical that overzealous use of overloading
broke overloading resolution in the compiler!). The version for TermRefs
would not consider `apply` insertion, but the version for Types would.
This means that the overloading broke the Liskov substitution principle. If
a general type happened to be a TermRef the more specific isApplicable
version would kick in, which however did less than the original.