Avoid spurious type test warning
This warning should trigger only for explicitly written type tests,
never for type tests coming from a match.

So far we did not notice this, since it triggered only for types that are
known to be null, and there are very few of those so far (basically,
just the primitive value types) and those types did not tend to be
tested irrefutably in patterns.