Handle differences in package vs package object in TypeComparer
With the changes in Typer, it is now possible that a type `T` that is a
member of some package object `pobj` in a package `p` is seen once as
`p.pobj.T` and then also as `p.T`. Type comparer has to make sure that
these two types are seen as equivalent.