Un-deprecate `option2Iterable`
The selection of methods available on `Option` is not very consistent.
For example, there is `.toList`, but not `.to` or any of the other
standard conversion methods like `.toVector`, `.toSet`, etc. There are
likely other methods that are commonly used today, but only available
through `option2Iterable` (e.g., `mkString`).

It's too late for 2.13 to go over Option's API. Keeping
`option2Iterable` covers up some of the shortcomings, and it's the same
as in 2.12, therefore backwards compatible.