Fix pickling positions problem
Make parameter definitions in DefDef have the same span as the symbols
they define. Previously, this position was not set, which means it would
be set pretty much in an ad-hoc way based on the positions that were set
in the definitions environment. For some reason I cannot quite reconstruct,
the changes to opaque caused pos/exports.scala to fail pickling tests
because an extension parameter symbol symbol a different position after
pickling.