Widen implied objects to their parent types
If we have a situation like this:

```
type A
type B <: A
implied a for A
implied b for B
```
we should recognize `b` to be more specific than `a`. Before this
commit that was not the case, since `a: A$`, `b: B$` and `B$` is not
a subtype of `A$`.